Move LSP Linearization impl to lsp crate
