+++ lsp/nls/src/linearization.rs
@@ -1,0 +1,206 @@
+use std::collections::HashMap;
+
+use nickel::{
+    identifier::Ident,
+    position::TermPos,
+    term::Term,
+    typecheck::{
+        linearization::{
+            Building, Completed, Environment, Linearization, LinearizationItem, Linearizer,
+            ScopeId, TermKind, Unresolved,
+        },
+        reporting::{to_type, NameReg},
+        TypeWrapper, UnifTable,
+    },
+};
+
+#[derive(Default)]
+pub struct BuildingResource {
+    pub linearization: Vec<LinearizationItem<Unresolved>>,
+    pub scope: HashMap<Vec<ScopeId>, Vec<usize>>,
+}
+
+trait BuildingExt {
+    fn push(&mut self, item: LinearizationItem<Unresolved>);
+    fn add_usage(&mut self, decl: usize, usage: usize);
+}
+
+impl BuildingExt for Linearization<Building<BuildingResource>> {
+    fn push(&mut self, item: LinearizationItem<Unresolved>) {
+        self.state
+            .resource
+            .scope
+            .remove(&item.scope)
+            .map(|mut s| {
+                s.push(item.id);
+                s
+            })
+            .or_else(|| Some(vec![item.id]))
+            .into_iter()
+            .for_each(|l| {
+                self.state.resource.scope.insert(item.scope.clone(), l);
+            });
+        self.state.resource.linearization.push(item);
+    }
+
+    fn add_usage(&mut self, decl: usize, usage: usize) {
+        match self
+            .state
+            .resource
+            .linearization
+            .get_mut(decl)
+            .expect("Coundt find parent")
+            .kind
+        {
+            TermKind::Structure => unreachable!(),
+            TermKind::Usage(_) => unreachable!(),
+            TermKind::Record(_) => unreachable!(),
+            TermKind::Declaration(_, ref mut usages) => usages.push(usage),
+        };
+    }
+}
+
+/// [Linearizer] used by the LSP
+///
+/// Tracks a _scope stable_ environment managing variable ident
+/// resolution
+pub struct AnalysisHost {
+    env: Environment,
+    scope: Vec<ScopeId>,
+}
+
+impl AnalysisHost {
+    pub fn new() -> Self {
+        AnalysisHost {
+            env: Environment::new(),
+            scope: Vec::new(),
+        }
+    }
+}
+
+impl Linearizer<BuildingResource, (UnifTable, HashMap<usize, Ident>)> for AnalysisHost {
+    fn add_term(
+        &mut self,
+        lin: &mut Linearization<Building<BuildingResource>>,
+        term: &Term,
+        pos: TermPos,
+        ty: TypeWrapper,
+    ) {
+        if pos == TermPos::None {
+            eprintln!("{:?}", term);
+            return;
+        }
+        let id = lin.state.resource.linearization.len();
+        match term {
+            Term::Let(ident, _, _) => {
+                self.env
+                    .insert(ident.to_owned(), lin.state.resource.linearization.len());
+                lin.push(LinearizationItem {
+                    id,
+                    ty,
+                    pos,
+                    scope: self.scope.clone(),
+                    kind: TermKind::Declaration(ident.to_string(), Vec::new()),
+                });
+            }
+            Term::Var(ident) => {
+                let parent = self.env.get(ident);
+                lin.push(LinearizationItem {
+                    id,
+                    pos,
+                    ty,
+                    scope: self.scope.clone(),
+                    // id = parent: full let binding including the body
+                    // id = parent + 1: actual delcaration scope, i.e. _ = < definition >
+                    kind: TermKind::Usage(parent.map(|id| id + 1)),
+                });
+                if let Some(parent) = parent {
+                    lin.add_usage(parent, id);
+                }
+            }
+            Term::Record(_, attrs) | Term::RecRecord(_, _, attrs) => lin.push(LinearizationItem {
+                id,
+                pos,
+                ty,
+                kind: TermKind::Record(attrs.clone()),
+                scope: self.scope.clone(),
+            }),
+
+            _ => lin.push(LinearizationItem {
+                id,
+                pos,
+                ty,
+                scope: self.scope.clone(),
+                kind: TermKind::Structure,
+            }),
+        }
+    }
+
+    /// [Self::add_term] produces a depth first representation or the
+    /// traversed AST. This function indexes items by _source position_.
+    /// Elements are reorderd to allow efficient lookup of elemts by
+    /// their location in the source.
+    ///
+    /// Additionally, resolves concrete types for all items.
+    fn linearize(
+        self,
+        lin: Linearization<Building<BuildingResource>>,
+        (mut table, reported_names): (UnifTable, HashMap<usize, Ident>),
+    ) -> Linearization<Completed> {
+        let mut lin_ = lin.state.resource.linearization;
+        eprintln!("linearizing");
+        lin_.sort_by_key(|item| match item.pos {
+            TermPos::Original(span) => (span.src_id, span.start),
+            TermPos::Inherited(span) => (span.src_id, span.start),
+            TermPos::None => {
+                eprintln!("{:?}", item);
+
+                unreachable!()
+            }
+        });
+
+        let mut id_mapping = HashMap::new();
+        lin_.iter()
+            .enumerate()
+            .for_each(|(index, LinearizationItem { id, .. })| {
+                id_mapping.insert(*id, index);
+            });
+
+        let lin_ = lin_
+            .into_iter()
+            .map(
+                |LinearizationItem {
+                     id,
+                     pos,
+                     ty,
+                     kind,
+                     scope,
+                 }| LinearizationItem {
+                    ty: to_type(&table, &reported_names, &mut NameReg::new(), ty),
+                    id,
+                    pos,
+                    kind,
+                    scope,
+                },
+            )
+            .collect();
+
+        eprintln!("Linearized {:#?}", &lin_);
+
+        Linearization::completed(Completed {
+            lin: lin_,
+            id_mapping,
+            scope_mapping: lin.state.resource.scope,
+        })
+    }
+
+    fn scope(&self, scope_id: ScopeId) -> Self {
+        let mut scope = self.scope.clone();
+        scope.push(scope_id);
+
+        AnalysisHost {
+            scope,
+            env: self.env.clone(),
+        }
+    }
+}
